requirements.txt
matplotlib
numpy
tqdm
pandas
Pillow
torch
torchvision
transformers
torchinfo
nltk
